vterm: bump to upstream 83ab18f, migrate stubs to new gr layout
Upstream rewrote struct gr from {fg, bg} (8 bytes) to {fg, bg, ul, a}
(16 bytes): three dedicated color words plus an attribute word.
Color-mode encoding changed (mode 00 is now default; 01 is 16-color
with full 0..15 directly addressable; the old 8..15-as-disguised-256
hack is gone). The gr_attrb() macro is replaced by a_get(g.a, FIELD).
Width moved out of bg's top bits into a's bits 22..25, but we read
c->w from struct vterm_cell, so that's invisible to us.

vterm_stubs.c:
- gr_color_to_ocaml is now a 4-way switch on mode bits, mapping
  mode 0 -> Default, 1 -> Basic, 2 -> Color256, 3 -> TrueColor.
  Bright colors flow through as Basic 8..15, which Grid emits as
  SGR 90..97 (terminals with customized bright palettes now get
  the right color).
- gr_to_attr uses a_get(g.a, BOLD) etc. Grid.attr shape unchanged;
  italic / strikethrough / underline-style / ul color are tracked by
  vterm but discarded at this boundary until Grid.attr is extended
  in a follow-up commit.

No user-visible behavior change.

See docs/SGR_PASSTHROUGH_PLAN.md for the full rollout.

+++ lib/vterm/vterm.h
@@ -19,7 +19,7 @@ struct vterm_cell {
 struct cursor_pos {
   int on_screen;
   int x,y, w;
-  uchar attrb;
+  uint32 attrb;     /* gr.a — full 22-bit rendition state plus width bits */
   int mode;
   union {
     const struct cell *cell;
